% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moloc.abf.R
\name{approx.bf.z}
\alias{approx.bf.z}
\title{Internal function, approx.bf.z}
\usage{
approx.bf.z(z, f, type, N, s, suffix = NULL, prior_var)
}
\arguments{
\item{z}{z statistics}

\item{f}{MAF}

\item{type}{"quant" or "cc"}

\item{N}{sample size}

\item{s}{proportion of samples that are cases, ignored if type=="quant"}

\item{suffix}{suffix to append to column names of returned data.frame}
}
\value{
data.frame containing lABF and intermediate calculations
}
\description{
Internal function, approx.bf.z
}
\details{
Calculate approximate Bayes Factors
}
\author{
Claudia Giambartolomei
}
\keyword{internal}
