% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moloc.abf.R
\name{change_indels}
\alias{change_indels}
\title{change_indels}
\usage{
change_indels(data)
}
\arguments{
\item{data, }{a data frame with columns A1 and A2}
}
\value{
data frame
}
\description{
Change indel format to I/D
}
\seealso{
Other align functions: 
\code{\link{complement_snp}()},
\code{\link{match_alleles}()}
}
\author{
Claudia Giambartolomei
}
\concept{align functions}
\keyword{internal}
