% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moloc.abf.R
\name{complement_snp}
\alias{complement_snp}
\title{complement_snp}
\usage{
complement_snp(x)
}
\arguments{
\item{x}{character}
}
\value{
character, complement
}
\description{
Find complement SNP
}
\examples{
complement_snp("A")

}
\seealso{
Other align functions: 
\code{\link{change_indels}()},
\code{\link{match_alleles}()}
}
\author{
James Boocock
}
\concept{align functions}
\keyword{internal}
