% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.moloc.R
\name{fast.coloc}
\alias{fast.coloc}
\title{To run a conventional colocalization}
\usage{
fast.coloc(coloc.data, type.a = c("quant", "cc"), type.b = c("quant", "cc"))
}
\description{
To run a conventional colocalization
}
