% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.moloc.R
\name{fast.moloc}
\alias{fast.moloc}
\title{To run multi-trait colocalization}
\usage{
fast.moloc(
  moloc.data,
  type.a = c("quant", "cc"),
  type.b = c("quant", "cc"),
  type.c = c("quant", "cc")
)
}
\description{
To run multi-trait colocalization
}
