% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.qc.data.R
\name{format.data}
\alias{format.data}
\title{Format GWAS Summary data into}
\usage{
\method{format}{data}(
  data = NULL,
  data.path = NULL,
  TRAIT,
  RSID = "RSID",
  EA = "EA",
  NEA = "NEA",
  EAF = NULL,
  MAF = NULL,
  BETA = "BETA",
  SE = "SE",
  N = "N"
)
}
\description{
Format GWAS Summary data into
}
