% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moloc.abf.R
\name{get_combos}
\alias{get_combos}
\title{get_combos}
\usage{
get_combos(x)
}
\arguments{
\item{x}{names of each data frame}
}
\value{
character vector with all combinations of names
}
\description{
Internal function, get_combos
}
\details{
Find all possible combinations for the traits
}
\author{
Claudia Giambartolomei
}
\keyword{internal}
