% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.data.R
\name{harmonise.moloc.data}
\alias{harmonise.moloc.data}
\title{To harmonise across 3 GWAS summary data, primarily for multi-trait colocalization analysis.}
\usage{
harmonise.moloc.data(data.a, data.b, data.c)
}
\description{
To harmonise across 3 GWAS summary data, primarily for multi-trait colocalization analysis.
}
