% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moloc.abf.R
\name{moloc.abf}
\alias{moloc.abf}
\title{moloc.abf}
\usage{
moloc.abf(
  listData,
  prior_var = c(0.01, 0.1, 0.5),
  priors = "default",
  save.SNP.info = FALSE
)
}
\arguments{
\item{listData}{A list of data frames to be analyzed. Each data frame
must contain columns named SNP (the SNP name consistent across all the data frames),
BETA,
SE}

\item{prior_var}{Can either be "default" or a numeric vector with one or more numbers for the prior variance of the ABF.
If use "default", 0.2^2 will be used for case-controls study while 0.15^2 will be used for quantitative traits.
If use a numeric vector, Approximate Bayes Factors will be computed with each prior and averaged.
By default, prior_var = c(0.01, 0.1, 0.5) where ABF will be calculated and averaged at prior_var = 0.01, 0.1, and 0.5.}

\item{priors}{Can either be "default" or a numeric vector, the prior for one variant to be associated with 1 trait and with each additional trait}

\item{...}{parameters passed to \code{\link{adjust_bfs}}, \code{\link{config_coloc}}, \code{\link{snp_ppa}}}
}
\value{
A list of three elements:
First is a data frame with 4 variables: priors ('prior'), likelihoods ('sumbf') and
posteriors ('loglkl' and 'PPA') for each configuration;
Second is a number, the number of SNPs in common in the region;
Third is a data frame with 2 variables:
SNP with the best posterior for each scenario ('coloc_ppas'),
SNP name with the best posterior for each scenario ('best.snp.coloc').
}
\description{
Bayesian multiple trait colocalization analysis using list of data.frames
}
\details{
Runs \code{\link{adjust_bfs}}, \code{\link{config_coloc}}, \code{\link{snp_ppa}}
}
\examples{
moloc = moloc.abf(listData) # uses default priors
}
\author{
Siwei Wu
}
