% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.data.R
\name{extract.data}
\alias{extract.data}
\title{Fast extract a subset of the data set in reference to the ID provided.
The data set where the data will be extracted should be a .txt file separated by tab}
\usage{
\method{extract}{data}(ref.id, data.path, data.id.index = 1)
}
\description{
Fast extract a subset of the data set in reference to the ID provided.
The data set where the data will be extracted should be a .txt file separated by tab
}
