% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.coloc.R
\name{fast.coloc}
\alias{fast.coloc}
\title{fast.coloc
To run a conventional colocalization using 2 GWAS summary data}
\usage{
fast.coloc(
  coloc.data,
  type.a = c("quant", "cc"),
  type.b = c("quant", "cc"),
  sdY.a = NULL,
  sdY.b = NULL
)
}
\arguments{
\item{coloc.data}{The harmonised GWAS summary data.}

\item{type.a}{Type of outcome of the 1st GWAS. "quant" stands for continuous outcome while "cc" stands for case-control design.}

\item{type.b}{Type of outcome of the 2nd GWAS. "quant" stands for continuous outcome while "cc" stands for case-control design.}
}
\value{
Colocalization result as coloc.abf in package coloc
}
\description{
fast.coloc
To run a conventional colocalization using 2 GWAS summary data
}
